Improve numeric stability for LayerNorm on CPU
Improve numeric stability for LayerNorm on CPU.

This diff will slightly decrease the performance of LayerNorm but increase the numerical stability especially when the variance is small.

Example.

M = 1024, N = 32768
input =  torch.randn(M, N, dtype=torch.float) * 0.01 + 100.0

Previously in this case, the computed variance will be all 0s.
After this diff, the variance will be

[1.002566e-04, 9.923905e-05, 1.014871e-04, ..., 1.015831e-04, 1.000394e-04, 1.009022e-04]

This is is close to 1e-4 as we expected. The max relative difference compare to double result is 0.0005683.

Differential Revision: [D21993646](https://our.internmc.facebook.com/intern/diff/D21993646/)

ghstack-source-id: 106259881
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40307