Improve numeric stability for LayerNorm on CUDA
Improve numeric stability for LayerNorm on CUDA.

This diff will slightly decrease the performance of LayerNorm but increase the numerical stability especially when the variance is small.

Example.
M = 1024, N = 32768
input =  torch.randn(M, N, dtype=torch.float, device="cuda") * 0.01 + 100.0

Previously in this case, the computed variance is
[0.      , 0.000977, 0.      , ..., 0.      , 0.      , 0.      ]

After this diff, the variance will be
[1.002436e-04, 9.923753e-05, 1.014664e-04, ..., 1.015593e-04, 1.000437e-04, 1.009089e-04]

This is is close to 1e-4 as we expected. The max relative difference compare to double result is 5.69883169e-05.

Differential Revision: [D21993664](https://our.internmc.facebook.com/intern/diff/D21993664/)

ghstack-source-id: 106260530
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40308